Module kernel_hal::console

Console input and output.

Structs

ConsoleWinSize

The POSIX winsize structure.

Functions

console_read

Read buffer data from console (serial).

console_win_size

Returns the size information of the console, see ConsoleWinSize.

console_write_fmt

Writes formatted data into the serial, and the graphic console if it exists.

console_write_str

Writes a string slice into the serial, and the graphic console if it exists.

debug_write_fmt

Writes formatted data into the serial through sbi call..

debug_write_str

Writes a string slice into the serial through sbi call.

graphic_console_write_fmt

Writes formatted data into the graphic console.

graphic_console_write_str

Writes a string slice into the graphic console.

serial_write_fmt

Writes formatted data into the serial.

serial_write_str

Writes a string slice into the serial.