Trait kernel_hal::drivers::scheme::EventScheme

pub trait EventScheme {
    type Event;
    fn trigger(&self, event: Self::Event);

    fn subscribe(
        &self, 
        handler: Box<dyn Fn(&Self::Event) + Sync + Send + 'static, Global>, 
        once: bool
    );
}

Associated Types

type Event

Required methods

fn trigger(&self, event: Self::Event)

Trigger the event manually and call its handler immediately.

fn subscribe(
&self,
handler: Box<dyn Fn(&Self::Event) + Sync + Send + 'static, Global>,
once: bool
)

Subscribe events, call the handler when an input event occurs. If once is ture, unsubscribe automatically after handling.