Module kernel_hal::interrupt

Functions

configure_irq

Configure the specified interrupt vector. If it is invoked, it muust be invoked prior to interrupt registration.

disable_irq

Disable IRQ.

enable_irq

Enable IRQ.

handle_irq

Handle IRQ.

is_valid_irq

Is a valid IRQ number.

msi_allocate_block

Method used for platform allocation of blocks of MSI and MSI-X compatible IRQ targets.

msi_free_block

Method used to free a block of MSI IRQs previously allocated by msi_alloc_block(). This does not unregister IRQ handlers.

msi_register_handler

Register a handler function for a given msi_id within an msi_block_t. Passing a NULL handler will effectively unregister a handler for a given msi_id within the block.

register_irq_handler

Add an interrupt handle to an IRQ

unregister_irq_handler

Remove the interrupt handle to an IRQ