Struct kernel_hal::UserContext

#[repr(C)]pub struct UserContext {
    pub general: GeneralRegs,
    pub trap_num: usize,
    pub error_code: usize,
}

User space context

Fields
general: GeneralRegstrap_num: usizeerror_code: usize
Methods

impl UserContext

pub fn run(&mut self)

Go to user space with the context, and come back when a trap occurs.

On return, the context will be reset to the status before the trap. Trap reason and error code will be placed at trap_num and error_code.

If the trap was triggered by syscall instruction, the trap_num will be set to 0x100.

If trap_num is 0x100, it will go user by sysret (rcx and r11 are dropped), otherwise it will use iret.

Example

use trapframe::{UserContext, GeneralRegs};

// init user space context
let mut context = UserContext {
    general: GeneralRegs {
        rip: 0x1000,
        rsp: 0x10000,
        ..Default::default()
    },
    ..Default::default()
};
// go to user
context.run();
// back from user
println!("back from user: {:#x?}", context);