Trait kernel_hal::PageTableTrait

pub trait PageTableTrait: Sync + Send {
    fn map(
        &mut self, 
        _vaddr: VirtAddr, 
        _paddr: PhysAddr, 
        _flags: MMUFlags
    ) -> Result<()>;

    fn unmap(&mut self, _vaddr: VirtAddr) -> Result<()>;

    fn protect(&mut self, _vaddr: VirtAddr, _flags: MMUFlags) -> Result<()>;

    fn query(&mut self, _vaddr: VirtAddr) -> Result<PhysAddr>;

    fn table_phys(&self) -> PhysAddr;

    fn map_many(
        &mut self, 
        vaddr: VirtAddr, 
        paddrs: &[PhysAddr], 
        flags: MMUFlags
    ) -> Result<()> { ... }

    fn map_cont(
        &mut self, 
        vaddr: VirtAddr, 
        paddr: PhysAddr, 
        pages: usize, 
        flags: MMUFlags
    ) -> Result<()> { ... }

    fn unmap_cont(&mut self, vaddr: VirtAddr, pages: usize) -> Result<()> { ... }
}

Required methods

fn map(
&mut self,
_vaddr: VirtAddr,
_paddr: PhysAddr,
_flags: MMUFlags
) -> Result<()>

Map the page of vaddr to the frame of paddr with flags.

fn unmap(&mut self, _vaddr: VirtAddr) -> Result<()>

Unmap the page of vaddr.

fn protect(&mut self, _vaddr: VirtAddr, _flags: MMUFlags) -> Result<()>

Change the flags of the page of vaddr.

fn query(&mut self, _vaddr: VirtAddr) -> Result<PhysAddr>

Query the physical address which the page of vaddr maps to.

fn table_phys(&self) -> PhysAddr

Get the physical address of root page table.

Loading content...
Provided methods

fn map_many(
&mut self,
vaddr: VirtAddr,
paddrs: &[PhysAddr],
flags: MMUFlags
) -> Result<()>

fn map_cont(
&mut self,
vaddr: VirtAddr,
paddr: PhysAddr,
pages: usize,
flags: MMUFlags
) -> Result<()>

fn unmap_cont(&mut self, vaddr: VirtAddr, pages: usize) -> Result<()>

Loading content...