Struct kernel_hal::user::UserPtr

#[repr(transparent)]
pub struct UserPtr<T, P: Policy>(_, _);

Raw pointer from user land.

Implementations

impl<T, P: Policy> UserPtr<T, P>

pub fn from_addr_size(addr: usize, size: usize) -> Result<Self, Error>

Checks if size is enough to save a value of T, then constructs a user pointer from its value addr.

pub fn is_null(&self) -> bool

Returns true if the pointer is null.

pub fn add(&self, count: usize) -> Self

Calculates the offset from a pointer. count is in units of T; e.g., a count of 3 represents a pointer offset of 3 * size_of::<T>() bytes.

pub fn as_addr(&self) -> VirtAddr

Returns the virtual address represented by the pointer.

pub fn check(&self) -> Result<(), Error>

Checks avaliability of the user pointer.

Returns [Ok(())] if it is neither null nor unaligned.

impl<T, P: Read> UserPtr<T, P>

pub fn as_ref(&self) -> &'static T

Converts to reference.

pub fn read(&self) -> Result<T, Error>

Reads the value from self without moving it. This leaves the memory in self unchanged.

pub fn read_if_not_null(&self) -> Result<Option<T>, Error>

Same as read, but returns None when pointer is null.

pub fn as_slice(&self, len: usize) -> Result<&'static [T], Error>

Forms a slice from a user pointer and a len.

pub fn read_array(&self, len: usize) -> Result<Vec<T>, Error>

Copies elements into a new Vec.

The len argument is the number of elements, not the number of bytes.

impl<P: Read> UserPtr<u8, P>

pub fn as_str(&self, len: usize) -> Result<&'static str, Error>

Forms an utf-8 string slice from a user pointer and a len.

pub fn as_c_str(&self) -> Result<&'static str, Error>

Forms a zero-terminated string slice from a user pointer to a c style string.

impl<P: 'static + Read> UserPtr<UserPtr<u8, P>, P>

pub fn read_cstring_array(&self) -> Result<Vec<String>, Error>

Copies a group of zero-terminated string into Strings, and collect them into a Vec.

impl<T, P: Write> UserPtr<T, P>

pub fn write(&mut self, value: T) -> Result<(), Error>

Overwrites a memory location with the given value without reading or dropping the old value.

pub fn write_if_not_null(&mut self, value: T) -> Result<(), Error>

Same as write, but does nothing and returns Ok when pointer is null.

pub fn write_array(&mut self, values: &[T]) -> Result<(), Error>

Copies values.len() * size_of<T> bytes from values to self. The source and destination may not overlap.

impl<P: Write> UserPtr<u8, P>

pub fn write_cstring(&mut self, s: &str) -> Result<(), Error>

Copies s to self, then write a '\0' for c style string.

impl<P: Policy> UserPtr<T, In>

pub fn read_iovecs(&self, count: usize) -> Result<IoVecs<P>, Error>