Module kernel_hal::interrupt

Interrupts management.

Functions

configure_irq

Configure the specified interrupt vector. If it is invoked, it must be invoked prior to interrupt registration.

handle_irq

Handle IRQ.

is_valid_irq

Is a valid IRQ number.

mask_irq

Disable IRQ.

msi_alloc_block

Method used for platform allocation of blocks of MSI and MSI-X compatible IRQ targets.

msi_free_block

Method used to free a block of MSI IRQs previously allocated by msi_alloc_block(). This does not unregister IRQ handlers.

msi_register_handler

Register a handler function for a given msi_id within an msi_block_t. Passing a NULL handler will effectively unregister a handler for a given msi_id within the block.

register_irq_handler

Add an interrupt handler to an IRQ.

unmask_irq

Enable IRQ.

unregister_irq_handler

Remove the interrupt handler to an IRQ.

wait_for_interrupt

Suspend the CPU (also enable interrupts) and wait for an interrupt to occurs, then disable interrupts.